Refactor webui interfaces into dedicated files
