+++ codegen/src/specs/webui/types/api-route-param.ts
@@ -1,0 +1,11 @@
+/**
+ * API route parameter definition
+ */
+export interface APIRouteParam {
+  name: string;
+  in: 'query' | 'path';
+  type: SchemaPrimitive;
+  required?: boolean;
+}
+
+export type SchemaPrimitive = 'string' | 'number' | 'boolean';
