+++ codegen/src/specs/webui/types/response-spec.ts
@@ -1,0 +1,10 @@
+import type { SchemaDefinition } from './schema-definition';
+
+/**
+ * API response specification
+ */
+export interface ResponseSpec {
+  status: number;
+  description?: string;
+  schema: SchemaDefinition;
+}
