+++ codegen/src/specs/webui/types/schema-definition.ts
@@ -1,0 +1,11 @@
+import type { SchemaPrimitive } from './api-route-param';
+
+/**
+ * Schema definition for request/response bodies
+ */
+export interface SchemaDefinition {
+  type: SchemaPrimitive | 'object' | 'array';
+  required?: boolean;
+  properties?: Record<string, SchemaDefinition>;
+  items?: SchemaDefinition;
+}
