+++ codegen/src/webui/components/types/search-filters.ts
@@ -1,0 +1,9 @@
+/**
+ * Filter options applied to search results
+ */
+export interface SearchFilters {
+  domain?: string;
+  platform?: string;
+  language?: string;
+  type?: string;
+}
