+++ codegen/src/webui/components/types/search-result.ts
@@ -1,0 +1,12 @@
+/**
+ * Represents a search result displayed in the generated UI
+ */
+export interface SearchResult {
+  id: string;
+  title: string;
+  summary: string;
+  domain: string;
+  type: string;
+  score: number;
+  highlights: string[];
+}
